refactored shelf into own component
